JDO-415 Three problems with the original test case are fixed with this patch.
1. As the bug noted, the DataStoreConnection is acquired outside a 
transaction, and there is no guarantee that the DataStoreConnection 
is even valid after beginning a transaction. The code now acquires 
the DataStoreConnection only after beginning a transaction. 

2. If the implementation supports Optimistic transactions, 
then the test should be repeated for both optimistic and 
datastore transactions. 

3. The transaction is never committed. This is fixed by putting 
the DataStoreConnection close and the transaction.commit into 
the finally block.